requirements.txt
affine==2.3.0
asgiref==3.3.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
bcrypt==3.2.0
bleach==3.3.1
bokeh==2.3.3
bson==0.5.10
Cartopy==0.19.0.post1
certifi==2020.12.5
cffi==1.14.5
channels==3.0.4
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.2
colorcet==2.0.6
common==0.1.2
constantly==15.1.0
cryptography==3.4.6
cycler==0.10.0
daphne==3.0.2
Django==3.1.7
django-crispy-forms==1.13.0
django-filter==2.4.0
djangorestframework==3.12.2
Fiona==1.8.20
Flask==1.1.2
forward==0.1.0
geographiclib==1.50
geopandas==0.9.0
geopy==2.1.0
geoviews==1.9.1
gepyto==0.10.1
h5py==3.2.1
holoviews==1.14.5
hvplot==0.7.3
hyperlink==21.0.0
idna==2.10
imageio==2.9.0
incremental==21.3.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
keras==2.6.0
kiwisolver==1.3.1
kneed==0.7.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.4.2
mglearn==0.1.9
movingpandas==0.7rc1
munch==2.5.0
nominatim==0.1
numpy==1.20.1
packaging==21.0
pandas==1.2.3
panel==0.12.0
param==1.11.1
paramiko==2.7.2
patsy==0.5.1
Pillow==8.3.1
pip-autoremove==0.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyct==0.4.8
pyfaidx==0.5.9.5
Pygments==2.8.1
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyplink==1.3.5
pyproj==3.1.0
pyshp==2.1.3
python-dateutil==2.8.1
pytz==2021.1
pyviz-comms==2.1.0
PyYAML==5.4.1
regexp==0.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.1
service-identity==21.1.0
Shapely==1.7.1
six==1.15.0
snuggs==1.4.7
SQLAlchemy==1.3.23
sqlparse==0.4.1
sshtunnel==0.4.0
statsmodels==0.12.2
threadpoolctl==2.2.0
tornado==6.1
tqdm==4.61.2
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.3
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
xlrd==2.0.1
zope.interface==5.4.0
